NAMING CONVENTION AND RESPONSOVE UI FIX
