+++ src/App.js
@@ -1,10 +1,11 @@
+import React from 'react'
 import './App.css';
-import Header from './header'
-import  MainContainer  from './mainContainer';
-import Footer from './footer';
+import Header from './Header'
+import  MainContainer  from './MainContainer';
+import Footer from './Footer';
 
 
-function App() {
+const App = () => {
   return (
     <div>
       <Header/>
