+++ src/Footer.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import './Header.css';
+
+const  Footer = () => {
+    return (
+      <div>
+      <div className = "footer">
+      <span style = {{paddingBottom:'10px'}}>Made in India</span>
+      </div>
+      </div>
+    );
+  }
+  export default Footer
+
