+++ src/ToggleTask.js
@@ -1,15 +1,16 @@
-import './mainContainer.css'
+import React from 'react'
+import './MainContainer.css'
 import previousBtn from './image/left-arrow.svg'
 import nextBtn from './image/right-arrow.svg'
 import addTasksBtn from './image/add-tasks.svg'
-import TaskCard from './taskCard';
+import TaskCard from './TaskCard';
 
 const ToggleTaskDays = () =>{
 return(
     <div className='toggle'>
-      <a className='previousTask'> <img src={previousBtn} alt="previous button" width="80" height="80" /></a>
-      <div className='taskDate'>Today's Task</div>
-      <a className='nextTask'> <img src={nextBtn} alt="next button" width="80" height="80"/></a>
+      <a className='previous-task'> <img src={previousBtn} alt="previous button" width="80" height="80" /></a>
+      <div className='task-date'>Today's Task</div>
+      <a className='next-task'> <img src={nextBtn} alt="next button" width="80" height="80"/></a>
     </div>
 );    
 }
@@ -26,7 +27,7 @@ const CompletedTaskLists = () =>{
         return(
         <div>
          <div style={{fontSize: '25px', display:'inline-block'}} > TO-DO:</div>
-           <a className= 'addTask'> <img src={addTasksBtn} alt="next button" width="40" height="40"/></a>
+           <a className= 'add-task'> <img src={addTasksBtn} alt="next button" width="40" height="40"/></a>
             <TaskCard title={"routine"} description="important tasks for the day" display/>
             <TaskCard title={"office meeting"} description="Attend daily scrum call @11am and present new requirements" display/>
          </div>
