--- src/footer.js
@@ -1,13 +1,0 @@
-import './header.css';
-
-function Footer() {
-    return (
-      <div>
-      <div className="footer">
-      <span style={{paddingBottom:'10px'}}>Made in India</span>
-      </div>
-      </div>
-    );
-  }
-  export default Footer
-
